polkavm-linker: properly handle relocations for dummy loads
Previously when we encountered a relocation for a load into the zero
register we just skipped handling that relocation, but this isn't
entirely right. Skipping the relocation will leave the original AUIPC
instruction as-is, which other parts of the code will complain about
since the code responsible for handling the relocations is supposed
to get rid of all of AUIPCs.

So here instead of skipping the handling just replace the AUIPC with
a NOP.