polkavm-linker: Add a return type to the import trampoline
and mark the return register properly.

We are seeing that in some debug builds trampoline subroutine is not inlined,
which causes epilogue and prologue to be generated. Further corrupting
the stack pointer and jumping to an incorrect address when returning from
the subroutine.

We fix this by adding a return type to the trampoline subroutine and marking
the return register properly, and therefore removing the "noreturn" attribute
from the trampoline subroutine.

Fixes: #277

Signed-off-by: Aman <aman@parity.io>