sandbox: Fix generic sandbox
We have been ignoring the generic sandbox for a while now. This commit
fixes the generic sandbox by implementing the updated Sandbox trait and
removing the old code.

Sandbox is used to provide isolation and interface to execute compiler
generated native code directly on the CPU. Generic Sandbox unlike Linux
Sandbox directly runs code inside Sandbox process by modifying process
memory and jumping directly to the guest code.

We implemented a bunch of interface but still some things are missing,
such as gas metering, aux data access and dynamic paging.

However, now we can use the generic sandbox for testing and it is able
to run 90% of our tests.

Signed-off-by: Aman <aman@parity.io>