sandbox: Fix gas metering for the generic sandbox
Generic sandbox uses a different base register, instead of r15 pointing
to Vmctx, it points to start of guest memory. Vmctx base address is
calculated by [r15-4096]. This increases the number of bits required for
offset when emitting instructions for gas metering.

As a result, we need more bytes for the gas metering instructions.

Following patch reworks gas metering logic to accommodate this change.

Signed-off-by: Aman <aman@parity.io>