linux-sandbox: Fix off-by-one errors in page_set accesses
We have off-by-one errors in the `page_set` accesses in the Linux sandbox
implementation. Current design requires `page_set`` to store inclusive
page ranges, but the code was incorrectly calculating the end of the range.

This commit fixes the off-by-one errors by adjusting the end of the page
range calculations to be inclusive.

Additionally, it adds a test to ensure that the memory access checks
work correctly after the fix.

Fixes: #308

Signed-off-by: Aman <aman@parity.io>