Interpreter: Allow reclaiming cache backed memory in the interpreter backend
This commit adds the ability to reclaim cache backed memory in the
interpreter backend of the PolkaVM. It deallocates compiled code and
frees up memory previously allocated by the backend. This is useful for
managing memory usage during the execution of multiple programs and
limiting memory footprint of the all the programs in the memory.

The obvious side effect of this change is that the cache is cleared and
the next program will have to be compiled again, which may lead to
slower execution for the first run after reclaiming memory.

Signed-off-by: Aman <aman@parity.io>