polkavm-compiler: Uninline division and remainder instructions
Instead of compiling individual division and remainder instructions inline,
we now compile them as calls to trampoline functions. This will reduce
the maximum number of compiled instruction per PolkaVM instruction.

We are doing this because division and remainder instructions are quite
rarely used in practice, and inlining them bloats the compiled code size
significantly.

Additionally, we have made division and remainder operations branchless.

Signed-off-by: Aman <aman@parity.io>