generic-sandbox: Fix invalid jump detection on macOS / Rosetta 2
On Rosetta 2, indirect jump to a non-canonical address doesn't trigger
a page fault (SIGBUS) right away. Instead, it jumps to the address, and then
triggers a page fault (SIGBUS) when trying to fetch the instruction.

This means that previous program counter is now lost, and now we have no
way of detecting if the guest program triggered an invalid jump or not.

We fix this program by updating the vmctx before making the jump and in
the signal handler we check if the current program counter is non-canonical.

Signed-off-by: Aman <aman@parity.io>